add auth with gates

+++ app/Http/Controllers/QuestionsController.php
@@ -6,6 +6,7 @@ use App\Http\Requests\AskQuestionRequest;
 use App\Question;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Gate;
 
 class QuestionsController extends Controller
 {
@@ -67,7 +68,11 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
-        return view('questions.edit', compact('question'));
+        if(Gate::allows('update-question', $question)){
+            return view('questions.edit', compact('question'));
+        }
+
+        abort(403, 'Access denied');
     }
 
     /**
@@ -79,6 +84,10 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if(Gate::allows('update-question', $question)){
+            return view('questions.edit', compact('question'));
+        }
+
         $question->update($request->only(['title', 'body']));
 
         return redirect()->route('questions.index')->with('success', 'Your questions has been updated.');
@@ -92,6 +101,10 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if(Gate::allows('delete-question', $question)){
+            return view('questions.edit', compact('question'));
+        }
+
         $question->delete();
 
         return redirect()->route('questions.index')->with('success', 'Your question has been deleted.');
